refactor trainer framework
