+++ towhee/trainer/metrics.py
@@ -1,0 +1,15 @@
+class Metrics():
+    def __init__(
+            self, metrics
+    ):
+        self.metrics = metrics
+
+    def add(self):
+        pass
+
+    def add_batch(self):
+        pass
+
+    def compute(self):
+        pass
+
