+++ towhee/trainer/modelcard.py
@@ -1,0 +1,28 @@
+import copy
+import json
+import os
+import warnings
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Union
+
+@dataclass
+class ModelCard:
+    # model_name: str
+    # finetuned_from: Optional[str] = None
+    # tasks: Optional[Union[str, List[str]]] = None
+    # dataset: Optional[Union[str, List[str]]] = None
+    # dataset_tags: Optional[Union[str, List[str]]] = None
+    # dataset_args: Optional[Union[str, List[str]]] = None
+    # eval_results: Optional[Dict[str, float]] = None
+    # eval_lines: Optional[List[str]] = None
+    # hyperparameters: Optional[Dict[str, Any]] = None
+
+    def __post_init__(self):
+        pass
+
+    def from_train_sumary(self):
+        pass
+
+    def from_train_config(self):
+        pass
